requirements.txt
streamlit
google-cloud-firestore
google-auth-oauthlib
streamlit_authenticator
bs4
requests
pandas
lxml
